l list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,17 +1,16 @@
-#include <stdlib.h>
 #include <stdio.h>
 #include "lists.h"
 
 /**
-  * print_list - Prints all elements of a list
-  *
-  * @h: A linked list
-  *
-  * Return: The number of nodes
-  */
+ * print_list - function with one argument
+ * @h: const pointer type
+ *
+ * Description: prints all the elements of a list_t list
+ * Return: number of nodes
+ */
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
+	unsigned int count = 0;
 
 	while (h != NULL)
 	{
@@ -19,11 +18,9 @@ size_t print_list(const list_t *h)
 			printf("[0] (nil)\n");
 		else
 			printf("[%d] %s\n", h->len, h->str);
-
 		h = h->next;
 		count++;
 	}
-
 	return (count);
 }
